+++ CHANGELOG.md
@@ -18,6 +18,8 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - ids parameter to transfer.query
 - hidden_fields parameter to Boleto.pdf
 - our_number attribute to Boleto
+### Fixed
+- Boleto.due convertion to datetime.datetime instead of misleading datetime.date
 
 ## [0.5.0] - 2020-08-11
 ### Added